requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
pyttsx3==2.90
SpeechRecognition==3.10.1
PyAudio==0.2.14
edge-tts==6.1.9
groq==0.4.2
python-dotenv==1.0.0
openai==1.12.0
numpy==1.26.2
soundfile==0.12.1
